requirements.txt
pyautogui
pywin32
qrcode[pil]
pillow
